fixup! 🐛 Fixed Offer Redemptions being over counted
